+++ integrations/langchain-py/.env.example
@@ -1,0 +1,2 @@
+BRAINTRUST_API_KEY=your-api-key-here
+OPENAI_API_KEY=your-openai-key-here
