+++ integrations/langchain-py/.envrc
@@ -1,0 +1,2 @@
+source_up_if_exists
+dotenv_if_exists
