+++ integrations/langchain-py/.tool-versions
@@ -1,0 +1,1 @@
+python 3.9.21
