+++ integrations/langchain-py/env.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+source venv/bin/activate
+export PYTHONPATH="$PWD/src:$PYTHONPATH"
