+++ integrations/langchain-py/pyrightconfig.json
@@ -1,0 +1,3 @@
+{
+  "typeCheckingMode": "strict"
+}
